gcp/requirements.txt
tensorflow
google-cloud-storage
Pillow
flask==2.2.5